requirements.txt
# Core dependencies
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Video processing
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
decord>=0.6.0
av>=10.0.0

# Deep learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Utilities
tqdm>=4.65.0
scipy>=1.10.0
scikit-learn>=1.3.0
pillow>=10.0.0
pyyaml>=6.0

# Annotation & visualization
pyqt5>=5.15.9
colorama>=0.4.6